PF/tfcbackend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
Pillow==9.3.0
geopy==2.2.0
stripe>=4.0.2
python-dateutil==2.8.2
django-crontab==0.7.1
django-override-autonow==0.0.1